+++ src/components/organisms/Post/HeroImage/HeroImage.style.ts
@@ -1,0 +1,19 @@
+import { GatsbyImage } from 'gatsby-plugin-image';
+import { css } from '@emotion/react';
+import styled from '@emotion/styled';
+
+const PostHeroImage = styled(GatsbyImage)`
+  ${({ theme }) => css`
+    width: 100%;
+    background-color: ${theme.color.primary};
+  `}
+`;
+
+const DefaultHeroImage = styled.img`
+  ${({ theme }) => css`
+    width: 100%;
+    background-color: ${theme.color.primary};
+  `}
+`;
+
+export { PostHeroImage, DefaultHeroImage };
