+++ src/components/organisms/Post/HeroImage/HeroImage.type.ts
