+++ src/components/organisms/Post/PostCard/PostCard.type.ts
