+++ src/components/organisms/Post/PostItem/PostItem.type.ts
