--- src/pages/post/components/index.ts
@@ -1,6 +1,0 @@
-import HeroImage from 'pages/post/components/HeroImage/HeroImage';
-import PostCard from 'pages/post/components/PostCard/PostCard';
-import PostItem from 'pages/post/components/PostItem/PostItem';
-import TagList from 'pages/post/components/TagList/TagList';
-
-export { HeroImage, PostCard, PostItem, TagList };
