+++ src/pages/post/{mdx.frontmatter__slug}.tsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 import { css } from '@emotion/react';
 import { DISPLAY } from 'styles/css';
 import { CenterLayout } from 'layouts';
-import { HeroImage, TagList } from 'pages/post/components';
+import { HeroImage, TagList } from 'components';
 import { usePostDetailPage } from 'pages/post/usePostDetailPage';
 
 const PostDetailPage = ({ data, children }) => {
